Rename history modules to event
Rename four history.rs files to event.rs and update all mod
declarations and use paths accordingly. No symbol names changed.
