+++ src/domain/entity.rs
@@ -2,5 +2,5 @@ pub mod author;
 pub mod book;
 pub mod common;
 pub mod event_set;
-pub mod history;
+pub mod event;
 pub mod user;
