+++ src/use_case/dto.rs
@@ -1,4 +1,4 @@
 pub mod author;
 pub mod book;
-pub mod history;
+pub mod event;
 pub mod user;
