+++ src/use_case/interactor.rs
@@ -1,6 +1,6 @@
 pub mod author;
 pub mod book;
-pub mod history;
+pub mod event;
 pub mod mutation;
 pub mod query;
 pub mod user;
