+++ src/use_case/traits/query.rs
@@ -7,7 +7,7 @@ use crate::use_case::{
     dto::{
         author::AuthorDto,
         book::BookDto,
-        history::{AuthorEventDto, BookEventDto},
+        event::{AuthorEventDto, BookEventDto},
         user::UserDto,
     },
     error::UseCaseError,
